Remove the IsOption type class
The function previously performed by the `IsOption` type class, namely,
the ability to control how a value of a specific type is rendered into
an `Options` value, is now taken care of by the specific `Option`
itself. `Option` now also admits a `Contravariant` instance due to this.

The behaviour previously captured by the `IsOption (Maybe a)` instance
is now captured slightly differently. If you had previously
written this:

    someOption := (value :: Maybe a)

you can get the old behaviour back (namely, the above expression
evaluating to an empty `Options` if `value` was `Nothing`) by changing
the code to this:

    maybe mempty (someOption :=) (value :: Maybe a)

There is also a newly provided helper function which captures the above
pattern: massoc, or its alias (:=?):

    someOption :=? (value :: Maybe a)

Most code should continue to work without modification, but note that
the above constitutes a breaking change (as, of course, does the removal
of `IsOption` and a few other related functions which are no longer
needed).

Bonuses:

* Everything is now documented
* There is no longer any use of the FFI